+++ ext_emconf.php
@@ -10,7 +10,7 @@ $EM_CONF[$_EXTKEY] = [
     'constraints' => [
         'depends' => [
             'php' => '8.2.0-0.0.0',
-            'typo3' => '12.4.0-13.4.99',
+            'typo3' => '13.4.6-13.4.99',
             'dashboard' => '',
         ],
         'conflicts' => [],