+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.0.0] - 2025-11-26
+
 ### Added
 - Compatibility with TYPO3 v14
 
@@ -254,7 +256,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release
 
 
-[Unreleased]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v3.4.0...HEAD
+[Unreleased]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v4.0.0...HEAD
+[4.0.0]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v3.4.0...v4.0.0
 [3.4.0]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v3.3.0...v3.4.0
 [3.3.0]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v3.2.3...v3.3.0
 [3.2.3]: https://github.com/brotkrueml/typo3-matomo-widgets/compare/v3.2.2...v3.2.3